+++ upcloud/resource_upcloud_network_peering_test.go
@@ -1,0 +1,58 @@
+package upcloud
+
+import (
+	"testing"
+
+	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/utils"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+)
+
+func peeringIDFunc(name string) func(state *terraform.State) (string, error) {
+	return func(state *terraform.State) (string, error) {
+		resourceState := state.RootModule().Resources[name]
+		return resourceState.Primary.ID, nil
+	}
+}
+
+func TestAccUpCloudNetworkPeering(t *testing.T) {
+	peering0 := "upcloud_network_peering.this.0"
+
+	testDataStep1 := utils.ReadTestDataFile(t, "testdata/upcloud_network_peering/network_peering_s1.tf")
+	testDataStep2 := utils.ReadTestDataFile(t, "testdata/upcloud_network_peering/network_peering_s2.tf")
+	testDataStep3 := utils.ReadTestDataFile(t, "testdata/upcloud_network_peering/network_peering_s3.tf")
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV5ProviderFactories: testAccProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testDataStep1,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(peering0, "name", "tf-acc-test-peering-peering-0-to-1"),
+					resource.TestCheckResourceAttr(peering0, "labels.%", "2"),
+				),
+			},
+			{
+				Config:            testDataStep1,
+				ResourceName:      "upcloud_network_peering.this[0]",
+				ImportState:       true,
+				ImportStateIdFunc: peeringIDFunc(peering0),
+				ImportStateVerify: true,
+			},
+			{
+				Config: testDataStep2,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(peering0, "labels.%", "1"),
+					resource.TestCheckResourceAttr("upcloud_network_peering.this.1", "labels.%", "1"),
+				),
+			},
+			{
+				Config: testDataStep3,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(peering0, "name", "tf-acc-test-peering-peering-0-to-1"),
+				),
+			},
+		},
+	})
+}